requirements.txt
streamlit
langchain
langchain_community
langchain-google-genai
google-generativeai
faiss-cpu
sentence-transformers
InstructorEmbedding
beautifulsoup4
pypdf
python-docx
python-dotenv